Use tinyvec's rustc_1_55 feature to reduce compilation time
While casually reading
https://hackmd.io/mxdn4U58Su-UQXwzOHpHag (Analysis of
rustc-benchmarking-data), I noticed that it mentions
https://github.com/Lokathor/tinyvec/issues/161.

The tinyvec crate, with the default set of features, can compile on
rustc < 1.55 by using large implementations of Default for [T; N].  On
Rustc >= 1.55, it can elide this code and just use const generics.
Apparently this reduces compilation time for tinyvec by about half -
which is not a lot in the context of librsvg, but let's make use of
that.

Librsvg currently requires Rust >= 1.64.0 anyway, so we can make use
of tinyvec's rustc_1_55 feature.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/822>