Copy a utility function to the librsvg-c tests to avoid having a test-utils feature
The only thing that librsvg-c's tests use from rsvg::test_utils is the
load_svg() function.  That function is trivial enough that we can just
copy it to librsvg-c's tests, and thus avoid having a whole feature
specification just for that.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/822>