Visual Studio: Further updates after separating crates
Make sure that we are building the correct target with Cargo when building the
internal Rust library for librsvg-c and update the intermediate target as
appropriate, so that we can properly link to it when building the final librsvg
DLL.

Also, reduce repetition in the NMake Makefiles by using Makefile variables
more.

Follow-up commit after changes in MR !822.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/824>