Add CI for Visual Studio
This is an initial stab at getting CI to work for Visual Studio builds,
where currently we attempt to build Pango and Gdk-Pixbuf first, with
pkg-config and libxml2 and FreeType standalone.  Pango builds do not
currently work  with subprojects building with FreeType, due to [1] as
FontConfig will be built anyways, so we build and install FreeType before we
build Pango.

After these, we will install Rust via rustup-init, and carry out the builds
for librsvg.

[1]: https://gitlab.freedesktop.org/cairo/cairo/-/merge_requests/469

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/826>