(#956): Rename rsvg-convert directory to rsvg_convert to avoid clashing with the rsvg-convert binary
When building with autotools and srcdir==builddir,
./target/.../rsvg-convert gets moved into ./rsvg-convert.  But since
./rsvg-convert is already a directory, we get an
./rsvg-convert/rsvg-convert binary.

Later, "make install" finds that ./rsvg-convert is a directory, not a
file, and fails to install the binary.

This renames the rsvg-convert directory to rsvg_convert, but makes
Cargo keep the name of the binary as rsvg-convert.  Another option
would be to let Cargo use rsvg_convert per the package/directory name,
and rename the resulting binary.

Thanks to Abderrahim Kitouni for figuring this out!

Fixes https://gitlab.gnome.org/GNOME/librsvg/-/issues/956

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/829>