CI: Actually test the build
We now expect librsvg to successfully build on the CI, as well as the C API
tests.

Not yet for Pango though, sadly, as upstream Cairo needs to be fixed, and we
run the Rust tests but do not expect it to pass, similar to what happens to the
MinGW/msys64 builds.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/831>