rust/Cargo.toml: Improve libxml2 dep lookup
Look also for "libxml2" for the package, so that the envvar that we set in the
NMake Makefile will work to pass in the libxml2 .lib.  Sadly, the
SYSTEM_DEPS_...() envvar will not work if our dependency name is just
'libxml-2.0', and we must use SYSTEM_DEPS_LIBXML2_NO_PKG_CONFIG as well.

libxml2 is required for all builds, so make it as such.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/831>