rsvg/Cargo.toml: Make FreeType compulsory for all builds
...and update the NMake Makefiles to feed in via envvars the FreeType .lib.
The reason why we are not using pkg-config files for Visual Studio builds is
because FreeType's build system support for generating pkg-config files is not
quite there yet, not unlike what is going on in libxml2.

For other notes, Cairo and HarfBuzz have pretty decent Meson build systems that
provide the pkg-config files we need, and at this point one should really not
be using Cairo-1.16.x.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/831>