StackingCtx: put the clip_rect from markers here, not in with_discrete_layer()
DrawingCtx::with_discrete_layer() takes a clip_rect argument, which is
an optional clipping rectangle in user space coordinates.  It is only
used from the markers code, because the markers code does not yet use
the clipping logic from the viewport code.

For now, I'll move that clip_rect into the StackingCtx passed to
with_discrete_layer(), with the intent that it should be the
StackingCtx that carries everything that needs to be known for drawing
a layer.  We can later refactor the markers code to do that using the
viewport code instead.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/838>