ci/env.sh: point CARGO_HOME to a local directory for local development
Normally the CI sets CARGO_HOME by itself to a directory that gitlab
can use for caching.  But for local development inside the container
image, CARGO_HOME will point to /root/.cargo, so that directory will
be deleted once the container exits.

Point it to a place in /srv/project so it persists between runs of the container.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/833>