NMake Makefiles: Allow specifying a previous stable Rust toolchain
Sadly, as found in issue #968, there may be cases where the latest stable Rust
toolchain installation isn't exactly that stable, or introduced changes in the
Rust syntax or so that breaks the build.  This adds an option TOOLCHAIN_VERSION
to the NMake Makefiles that forces a previous-stable version that is installed
to be used instead of the latest installed stable toolchain release.  Since the
toolchain must have been previously installed, Cargo will check for us the
validity of the version that is specified.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/837>