NMake Builds: Use rustc to query the required system libs
Extend the helper batch file that we use to query the default Rust toolchain so
that we can use the rustc that corresponds to the Rust toolchain that we are
using to query the system libraries (i.e. from the Windows SDK), in order to
ensure that the final librvg DLL links, as Rust might involve changes that
require more system libraries to be required for the build, such as in the
issue below:

https://github.com/rust-lang/rust/pull/108262

Fixes issue #968 for the librsvg-2.56 branch.