NMake: Add warning messages about stub python.exe
Windows 10/11 is likely to default to a stub python.exe, which was initially
intended for people to install Python from the Microsoft Store, which may
get into our way if one does not specify PYTHON to point at a real Python
installation. So, show messages where relevant if a stub python.exe is found
in one's %LOCALAPPDATA% and PYTHON is not specified.

Also drop an extraneous include to introspection-msvc.mak in Makefile.vc.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/847>