(#998): Fix cascade for symbol elements being referenced from use elements.
In ffbec91a there was an inadvertent renaming of a variable,
"symbol_values" -> "values", and "values" would then shadow the main
ComputedValues from the <use> element being rendered.  In the fix I'm
calling it "child_values" to be consistent with the rest of the
function.

Fixes https://gitlab.gnome.org/GNOME/librsvg/-/issues/992