CI: Run cargo-doc in a separate job from "pages"
The pages job only runs after a merge, but cargo-doc may actually fail
if there are errors in the documentation process.  This happened with
the last merge since the pages job was still calling cargo-doc with
the old project name (librsvg instead of rsvg).

We'll fix the latter in the next commit.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/859>