Remove unnecessary byteorder dependency
There is a single place in the code base using byteorder. This writes a
pixel value as u32 in native endian into a byte array aka [u8].

Instead of using byteorder for this, it seems simpler to use
u32::to_ne_bytes() to get the bytes and copy_from_slice() to do the
copy.

Signed-off-by: Uli Schlachter <psychon@znc.in>
Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/876>