gen-rust-docs.sh: don't pollute the rsvg public docs with links to the internal crates
We build the internals and public docs, but both get written to
target/doc - this causes the final result to have the public docs for
the rsvg crate also contain links to the internal crates.

This happens because we don't clean up target/doc/ after copying the
internals docs to their final public/internals destination.

Now we move the files instead, and then generate the public docs.
This gives us a clean target/doc for the public docs to be generated
on, and so the final public/doc doesn't have links to crates that are
internal only.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/878>