Do not expose AcceptLanguageError in the public API
This is one of those errors with too much detail, and the calling
program cannot really do anything with it except pass an actually
valid Accept-Language header payload to AcceptLanguage::parse().

So, just use a String for that function's error case.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/878>