Add C API tests for calling the API out of order
E.g. starting to load a handle, then trying to render it while loading
is not yet complete.

Why are we not getting coverage info for this (e.g. the errors in
librsvg-c/src/handle.rs:get_handle_ref)?  Is it because the g_test
subprocess used to capture the g_critical message?

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/889>