CI: Fix vs2017-x64 CI's pkg-config build
We need to use `md` instead of `mkdir` in pkg-config's NMake Makefiles as GNU's
version might be picked up instead if the GNU tools are in the system's %PATH%,
which is not what we want, and we can't override the %PATH% for forcing the
Windows' version of `mkdir` it since it is a built-in command in Windows cmd.exe
shell.  Also, for completeness' sake, use `rd` instead of `rmdir`, for similar
reasoning.

Since this involves updating pkg-config's NMake Makfiles that is ideally done
upstream, but it seems that it has not been updated for a while, patch
pkg-config's NMake Makefiles instead to fix this.

As a result, expect the vs2017-x64 CI job to pass again.

Fixes issue #1020.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/895>