Use Rc<PaintSource> instead of Arc
On Rust 1.72, clippy complains that Arc is being used around something
that is !Send and !Sync.  We could turn off that warning, but...

ResolvedPattern holds a rctree::WeakNode, which is !Send and !Sync
since it refers to an Rc.

At some point we'll have to switch to an Arc-based tree of nodes, or
to numerical ids or something.  For now, use Rc<PaintSource>
everywhere instead.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/896>