(#1025): ci/setup-dependencies-env.sh: Extract script to set up PATH and such for a group of dependencies
The current instructions for users of the ci/pull-container-image.sh script are
incomplete, since we now require PATH/LD_LIBRARY_PATH/PKG_CONFIG_PATH
to be set with respect to /usr/local/librsvg, which is where the
libraries for dependencies are installed.

This new script can be sourced into the current shell to adjust the
environment as appropriate; it uses the PREFIX variable as a basis.
The prefix path is not hardcoded so that the same script can be used
if we install the dependencies in a different prefix, for example with
different compilation options for fuzzing.

Fixes https://gitlab.gnome.org/GNOME/librsvg/-/issues/1025

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/898>