Reference::from_png(): make infallible
This struct, and its functions, are only meant to be used from the
test suite.  In that case we can take the luxury of not returning a
Result, and to just expect() that e.g. reference PNGs can be read
correctly.

(This also lets me figure out what is going on when I add a new test
SVG, but I forget to add the corresponding reference PNG - now the
failure prints out the offending filename.)

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/901>