Use preserveAspectRatio=none when rendering a toplevel SVG from an <image> element
This commit introduces an SvgNesting enum, which indicates whether a
toplevel SVG is being rendered Standalone, or
ReferencedFromImageElement (e.g. `<image href="foo.svg"/>`).

In the second case, the nested SVG needs to use
preserveAspectRatio=none so that it will scale all the way to the
viewport defined by the parent document's <image> element.  This is
somewhat buried in the spec; the code has a comment with the relevant
URL.

We pass this enum from the toplevel handle::render_document(),
used by the implementation of the <image> element, down to DrawingCtx.
Then, structure::Svg can access it and see what sort of
preserveAspectRatio to use: its own as normal, or `none` if it is
being referenced from an <image> element.

This commit also removes debugging paraphernalia that I put in for
this bug, and updates the reference image for
coords-viewattr-04-f-ref.png now that it renders correctly.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/901>