CI: allow the "check devel-docs links" job to fail if not in the main branch
The devel docs are only published for the main branch anyway, and I
just ran into the following:

I removed rsvg/src/handle.rs in the main branch, but the devel-docs in
librsvg-2.57 still refer to it.  Which makes sense, since in *that*
branch handle.rs still exists - even though the devel-docs for old
branches are never published.

So, allow a soft failure for this job if it's not for the main
branch.  Hopefully it will still catch broken links to external sites.