Add a SharedImageSurface::from_image() method
This converts from an image::DynamicImage to an ARGB
SharedImageSurface, as usual.

The only new thing needed to support this is to
"impl ToPixel for image::Rgba<u8>" - the rest of the code already uses
ToPixel to convert pixel formats, so we can just cut&paste the
code to convert the whole image with a couple of iterators.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/904>