Meson: Use --release for Cargo for debugoptimized MSVC-style builds
With the Visual Studio Rust toolchain, .pdb files are also generated when the
'--release' flag is used for Cargo, and using that flag fits the purpose of
'debugoptimized' better.

Also copy and install the .pdb file for rsvg-convert when debug symbols are
requested for the build.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/932>