qdk-pixbuf-loader: Install the loaders relative to the destdir if set
We pass the paths we from pkgconfig to the loader install
script. However if we have a destdir setup, we should
modify the paths so we install in the destdir rather than
the directly on the path we got from pkgconfig.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/933>