meson/*.py: Fix running on Windows with Python-3.7.x
subprocess.run() doesn't like Path's that are WindowsPath's in the
command lines, so make sure things are applied with .as_posix() so that
subprocess.run() accepts things on Python 3.7.x.

This is done as the required Meson version still supports Python
3.7.x.

Note that this does not cover MinGW/mingw-w64 builds, as posix'ish paths
are being used there.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/949>