meson: Windows: Check RustC for default target triplet
Update the meson/query-rustc.py script to also allow querying RustC
for the default target triplet that is currently selected.

If the default target triplet matches what our build is for, or if
-Dtriplet=... is specified, proceed as before; otherwise, construct a
--target=... argument that best matches our build to feed into Cargo.

This way, this reduces the likelihood that `-Dtriplet` needs to be used,
and reducing getting into build errors because the default RustC
target triplet did not match our build target.

Check also whether the default (or requested) target triplet exists for
the current build toolchain.

This is currently not done for non-Windows platforms as I am not
well-versed enough about how things go on these platforms.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/936>