meson: Fix generating static and both libraries under MSVC
Back when I implemented the mesonification of librsvg, I forced the
installable target of rsvg-2 to be a shared library. This was because
Meson and Rust, at link time, differ on the naming convention expected
of static libraries:

https://github.com/rust-lang/rust/issues/122455

https://mesonbuild.com/FAQ.html#why-does-building-my-project-with-msvc-output-static-libraries-called-libfooa

I did not realise at the time that the fix merged in !938 (commit
73d71672750f397e7b5311e385ebf020f90abb98) was not aware of this issue,
and so it broke the Cerbero builds targeting Microsoft compilers.
This commit works around the Rust bug by following the Linux convention for
static libraries, and then using a custom target to rename the Rust-generated
library if under MSVC.

All targets that depend on librsvg_lib do so now unconditionally, given
they must link at a time where the file outputs are settled.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/957>