Map 'serif' to 'sans-serif' since we don't have a serif font yet
Two things:

* A couple of test files use "serif" for the font-family name.  Since
we don't have such a font, or a mapping, fontconfig tries to match on
the rest of the properties.

* So, when a test file requests "serif Bold", it goes to DejaVu Sans
Bold.  We don't want this.

So, explicitly map "serif" to "sans-serif" so it gets resolved to
Roboto eventually.

At some point I may add an actual serif font for testing purposes.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/965>