(#1080): Document that the default DPI for RsvgHandle is 90
For historical reasons, it was kept at 90, and must remain so to avoid
changing the behavior in applications.  This commit makes a
recommendation in the docs to change the DPI to 96 before rendering.

Fixes https://gitlab.gnome.org/GNOME/librsvg/-/issues/1080

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/969>