build: Disallow introspection on Windows static-only builds
Introspection on Windows look through the built .lib file fed to it to see
which DLL it is linked to, so that it fills in the "shared-library" entry in
the generated .gir file.  Introspection is currently not supported on static
library builds on Windows.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/968>