build: Fix introspection on Visual Studio-like builds
...when default_library='both'.

Use the librsvg_shared_lib target for the shared library that introspection
will use.  This is because:

 * The 'rsvg-gir' target used for introspection when default_library='both'
   does not generate a .lib file when the DLL is built, on Visual Studio-like
   compilers.
 * More importantly, the 'rsvg-gir.dll' that results is not installed, and
   libgirepository will not be happy when 'Rsvg-2.0.typelib' is requested, as
   libgirepository will attempt (unsuccessfully) to load 'rsvg-gir.dll'.

This might affect MinGW/mingw-w64 or even other builds as well, but I am not
knowledgeable enough in that toolchain to say anything about that, so I am
currently leaving that alone.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/968>