Try to propagate an error earlier when there are circular references
The problematic file has circular references, and they are detected,
but the <use> element tries to just return an empty bbox rather than
propagate the error upstream.  The problem with this particular file
is not that it enters infinite recursion, but that the <g> are nested
so deep that we run out of stack.

So, propagate errors about circular references upstream as much as
possible, and stop propagating when we reach Node::draw().

The effect now is that rendering proceeds until the maximum number of
referenced elements is reached, and then we get a hard error that
makes the rendering terminate.

I'm not 100% satisfied with this approach; it fixes the stack
exhaustion, but we could try other things:

- Mark <use> elements as "has a circular reference" so they are not
re-rendered if referenced again.

- Mark all the elements involved in a circular reference chain as
tainted, so they can't get referenced again.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/988>