meson/cargo_wrapper.py: Improve Windows check
The original check for whether we are building for Windows so to copy the
rsvg-convert utility correctly is only sufficient when the build is done on
Windows, but did not account for cross builds done on other systems targetting
Windows.

This improves the check to do the following:

* If no --target argument is specified, check whether the build is done on
  Windows, as we did before.
* If a --target argument is specified, see whether it contains 'windows'
  in the <arch>-<platform>-<os>-<compiler_toolchain> that is specified,
  specifically in the <os> field, as all supported Windows targets at [1]
  go in this pattern.

We currently do not account for custom Windows targets at the moment.

Should fix issue #1084.

[1]: https://doc.rust-lang.org/nightly/rustc/platform-support.html

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/994>