rsvg/tests: Replace failing ref PNGs
There are slight differences for how the image is rendered between
image-rs 0.24 and 0.25, which can happen since different JPEG decoders don't
deliver the same pixel perfect results.

The reason why this happens with image-rs 0.25 is that it uses a
completely different decoder (zune-jpeg) which seems to be more correct
and very fast.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/1001>