(#1102): Remove obsolete test for Bugzilla bug 749415
https://bugzilla.gnome.org/show_bug.cgi?id=749415 was caused by
incorrect and duplicated logic to decide when to create a temporary
surface for a stacking context (... back when we didn't even have the
concept of a stacking context).

The test is a chunk of a diagram with a bunch of text elements with
very particular transforms.  Any minute change in the text rendering
stack tends to change the rendering slightly.

This is not a useful test anymore; we already have other tests for
text with transforms.

For reference, the commit that originally introduced the tests is
cb7fd6b635a519e968ea67ec529651fe774abe26

Fixes https://gitlab.gnome.org/GNOME/librsvg/-/issues/1102

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/1009>