pixbufloader: fix missing prefix relocation for the installation paths
In case the install prefix and the gdk-pixbuf prefix are different (which
is the default in MSYS2 packaging for example) we need to override the
prefix defined in the .pc file with the one we are installing to.

This fixes the loader being installed into the wrong directory under MSYS2.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/1015>