visual Studio builds: Fix introspection on latest Windows SDK
The latest Windows SDK (10.0.26100.0) enforces that check symbols are
built into binaries that are using it, where the CRT cflag (/MD or /MDd
and so on) determines the appropriate symbol that is needed at link time.

This means, the introspection build fails as we were not specifying a
CRT cflag when we run g-ir-scanner; so this attempts to remedy the
situation by also passing in the appropriate CRT cflag during the
introspection build.