layout::Path - Use an enum to differentiate between validated and unvalidated paths
As a mitigation for Cairo's bugs, we will try to keep from feeding it
paths that overflow Cairo's representable range of coordinates.

For now we just have a Validated variant; the next commit will
introduce the validation and the invalid variant.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/1029>