Add a test file for #1118, from fuzz testing
This is a reduced version of the original fuzzed SVG posted to the
bug.  The steps I took to reduce it are as follows - after each step I
ran the crashing test again to see if it still crashed in the same
way.  I stopped making changes when the crash no longer happened, and
left the last state of the SVG that still caused a crash.

* remove duplicated element ids - librsvg only picks up the first one
  of each.

* remove elements inside <defs> that ended up without ids, since those
  cannot be referenced anyway.

* remove graphic elements that experimentally did not have any effect
  on the crash.

* remove masks without children.

* remove masks that experimentally did not have any effect on the
crash.

* remove masks without ids since they cannot be referenced.

* add ids to all the <use> elements to see which is which

* remove empty groups

* promote elements up to the grandparent group when the parent group
  has no attributes, just the raw <g>

* remove fill attributes in use elements

* remove matrix transform that experimentally did not have any effect
  on the crash.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/1030>