Annotate places where the transforms need to be kept in sync
For now I'll leave these as placeholders: the viewport.transform is
just bookkeeping that is not used anywhere except for
Path::has_unsuitable_coordinates().  And the way it is used *just
there* currently makes the library not crash with the fuzzed file
anymore.

To *really* keep the transform in the viewport in sync with the Cairo
transform, I'll fix these "FMQ" comments shortly, after doing a
release with the immediate mitigation for the fuzzing crash.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/1030>