Remove the deep-nesting test for now
It still overflows its stack on the CI.  I don't know if the CI is
using a smaller thread stack size (?) than when running with the CI
container on my machine.

So, for now, we'll remove it, and leave its ghost in version control
for when I'm too lazy to rewrite it by hand later.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/1038>