rsvg-bench.rst: Remove sections that are superfluous now that the code is in this repo
When rsvg-bench was a standalone tool in its own repository, it made
sense to link to the code of conduct and the maintainer's details from
its documentation.

Now that rsvg-bench is contained right here, those sections are
available elsewhere in the source tree.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/1039>