query-rustc: fix running with rust 1.82
Starting with rust 1.82 the internal rustc call fails with:

failed to rename archive file: Access is denied. (os error 5)

For some reason rust doesn't like Python having the file open while
it tries to write to it (or replace it). See upstream issue:
https://github.com/rust-lang/rust/issues/131896

Work around by creating a temporary file and closing the handle before
calling into rust, and deleting the file at the end.

This could also be done with NamedTemporaryFile() by passing
delete_on_close=False to it and using it as a context manager,
but that only works since Python 3.12.

Fixes #1134

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/1040>