CairoPath: introduce our own representation of a Cairo path
This is just built from cairo::PathSegment.  Instead of having a
path.to_cairo(cr) directly modify the cr, we'll have an intermediate
representation with CairoPath.

In the following commits first we'll lower a path_builder::Path to a
CairoPath and validate its coordiantes.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/1053>