(#1109): Fix text centering for scaled text
Use an identity transform for text layout, instead of using the
transform that the <text> element would use within the current
stacking context.

It used to be that the text layout code required knowing the final
transform of the current stacking context, to pass it to Pango,
because of rendering differences when rendering scaled text
vs. rendering text and *then* scaling it.

However, those differences only happen when Pango is in its default
mode of using outline hinting for the glyphs.

Librsvg explicitly turns off hinting for text rendering, since that is
the expectation for SVG rendering.  So, we can actually just use an
identity transform when laying out text in Cairo, and the result will
be scaled/rotated to its final position just fine.

Fixes https://gitlab.gnome.org/GNOME/librsvg/-/issues/1109

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/1047>