meson: remove MSYS2 specific workaround for rust target selection
In MSYS2 we used the gnu target even for llvm envs to work around
the rust ecosystem not supporting that target.

A month ago we switched to gnullvm, see
https://github.com/msys2/MINGW-packages/pull/21902

Remove the MSYS2 workaround to fix the build in MSYS2 llvm envs.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/1052>