(#1138): Report errors correctly from the parsers for attribute values
Say we have an attribute with a CSS <number> type:

  bar="42"

Due to the way attributes were parsed, as QualName.parse("42"), it was
possible to have

  bar="42 43"

so that only the first number would be used, and the rest ignored.
This is because "impl<T: Parse> ParseValue<T> for QualName" did not
check for token exhaustion - it worked in the same way as property
parsers.

Unlike property values, attribute values must parse the entire string,
since attribute parsers are not composable into larger parsers.

In this particular bug, we had the following incorrect value from a
`values` attribute for a color matrix:

      <feColorMatrix
         values="1 0 0 0 0 0 1 0 0 0 0 0 1 0 0 0 0 0 0,500000 0 "

That attribute expects a list of 20 numbers separated by whitespace or
commas, but here we have 21 values (it's a bug in an old version of
Inkscape that wrote out floats with the locale separator for
decimals).

So, the parser for 20 matrix values worked correctly up to the 500000,
and then the final 0 was ignored.  This is what caused the overall
parser for that attribute value to indicate that there was no error.

Now, we actually call parser.expect_exhausted() when parsing an
attribute value.  This causes the error to be reported correctly.

Fixes https://gitlab.gnome.org/GNOME/librsvg/-/issues/1138

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/1063>