Parse the x/y/dx/dy attributes of <text> with a CommaSeparatedList
Per the limitation from bug #183, we still ignore all but the first
value in each of those attributes.  However, this change (explicitly
parsing lists of length values) will let us change the general
function to parse attribute values to check for token exhaustion.

Right now x/y/dx/dy with multiple values *can* be parsed, but only
because that general function does not check for token exhaustion.  So
it parses the first Length, and ignore anything else in the value
string.  We'll change that to be strict and parse the whole thing, but
for that we need this change.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/1063>