First homework: Add a test for collecting text content from a node into a a single string
If we have a text element like

  <text>
    foo
    <tspan>bar</tspan>
  </text>

We need to be able to gather all of its text content into a single
string, "foo bar" (plus extra whitespace from around the XML tags).

This is the first test to make sure that happens correctly.

Anointing will implement the code to do that.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/1100>