Extract get_transform() from DrawingCtx and have it explicitly take a cr
This makes it very clear that the transform is stored in the cr, and
it shouldn't be where we maintain it.

We'll use this function in the masking code, since that one needs the
transform from the mask_cr, which is not the DrawingCtx's cr.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/1099>