Explicitly pass clip_to_rectangle() the transform to use
This is redundant right now, since that function assumed that it would
be using the passed cr's current transform, and we *are* now just
passing that cr's transform.

However, I want to make it more clear what parts of the code assume
the current transform in the cr, so that they can be made to use the
transform from the viewport instead.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/1099>