Remove redundant condition
We use rect_to_transform() now, and that function takes the
CoordUnits, *and* it returns an identity transform if the units are
userSpaceOnUse.  So we can just always transform the mask_rect with
the returned transform; there's no need to special-case userSpaceOnUse now.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/1099>