Use an explicit transform for the new viewport for the temporary surface
Now I understand that whenever we use DrawingCtx::nested(), to set up
the draw_ctx for a temporary surface, we also need to create a new
Viewport that starts with the special transform for temporary
surfaces.

There's a couple other places where this is not done yet; I'll get to them.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/1099>