cairo-draw: Convert pango layout to path and draw that
This is done to work arround a pango regression (pango#520) that causes
it to improperly render pango layouts in cairo contexts with a CTM that
has non uniform scaling factors (M = [m1 m2 t]; norm(m1) != norm(m2)).

See #599 for the discussion related to the pango regression in librsvg.

The mechanism this introduces is *roughly* similar to the one introduced
by 7fd70e1, with the added change that it takes a uniform scaling factor
from the CTM for use with pango to avoid issues with very small font
sizes combined with transformations with large scaling factors (Pango
will use the y-scale from the CTM to adjust its fontsize). Newer rsvg
seemingly wasn't affected by that for some reason.