styles: Apply styles at the end of parsing
Previously, CSS style nodes would only apply to new nodes parsed after
the style node. This doesn't match other SVG engines and also makes it
hard to implement loading additional stylesheets after the initial load.

Now, instead of applying the style during set_atts, we wait until the
whole document has been parsed, after which we iterate over the tree
and apply the latest styles to each node. The downside of making this
possible is the increased memory usage. This is due to having to store
all XML attributes of a node in the node, as we need them to reevaluate
its style.