path: Fixed parsing or arc flags without spaces
The elliptical arc command contains two special "flag" values which can
be either 0 or 1. The SVG path grammar special-cases those flags
compared to other numbers (which this rsvg assumed before), allowing
them to be placed next to each other without any required seperation.
They are still allowed to be written seperately with (\w*,\w*)? between
them (Note that this implementation doesn't fully match, instead it
allows for an abirary amount of commas between them, which should be a
mal-formed path).

The implementation of this was inspired by what emilesnyder did in
7b17b55.

Also added testcase for this, icon is taken from libadwaita (LGPL-2.1)