meson: build libcroco statically by default
For GTK, we'd like to statically link this rsvg version, that we don't
have to dump it into the libdir, which may cause breakage.
This kinda extends to libcroco aswell, altho this is primarily about not
instaling unnecessary libararies and not about breakage (no-one but us
should use it in the first place).