meson: hack to avoid symbol leakage for static builds
If you build a static librsvg-2.a and let meson link it into your shared
library, GNU style linkers will include all used librsvg-2.a symbols in
the exposed symbol table for the shared library. Outside of a few
exceptions, this is not what we want (if you do, you can probably get
away with partial_dependency to filter out the linker flags), so lets
pass -Wl,--exclude-libs librsvg-2.a to the linker, which causes it to
not include those symbols in the symbol table.

This should probably considered a *bug* in meson and future versions
might handle such flag-passing internally, but we'll do it manually for
now.

See https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8541#note_2462602