cairo-clip: fixed leak if testing set_testing was enabled
Previously, the clip renderer subclass did not chain up to the free
implementation of the cairo renderer, nor did it even call its free
function in the vtable, opting to use g_free directly.

This works, as long as you can actually free the renderer instance with
a simple g_free, which is the case unless you have called
    rsvg_handle_internal_set_testing (handle, TRUE)
as that'll result in the font_config_for_testing & font_map_for_testing
members getting allocated in the cairo renderer instance, which will
have to be explicitly freed.

This commit reworks the clip renderer, that it
1. chains up to the free call of the cairo renderer
2. actually uses rsvg_render_free to free itself instead of just calling
   g_free.